requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
uuid==1.30
